All Assets D
This includes the skyboxes, all of the weapon prefabs and codes (both for the player and the enemy), and the scripts and models that the game does not use. Include all assets in the Assets folder before attempting to run!
Note: If you want to see what currently runs on the Quest, look in the Procedural Asteroid folder for the level scene.